Filter out warnings inside `check_crate`
