+++ tooling/nargo/src/ops/compile.rs
@@ -1,10 +1,8 @@
-use std::collections::HashSet;
-
-use fm::{FileId, FileManager};
+use fm::FileManager;
 use noirc_driver::{
     link_to_debug_crate, CompilationResult, CompileOptions, CompiledContract, CompiledProgram,
 };
-use noirc_errors::FileDiagnostic;
+
 use noirc_frontend::debug::DebugInstrumenter;
 use noirc_frontend::hir::ParsedFiles;
 
@@ -135,12 +133,10 @@ pub fn collect_errors<T>(results: Vec<CompilationResult<T>>) -> CompilationResul
 pub fn report_errors<T>(
     result: CompilationResult<T>,
     file_manager: &FileManager,
-    root_files: &HashSet<FileId>,
     deny_warnings: bool,
     silence_warnings: bool,
 ) -> Result<T, CompileError> {
     let (t, errors) = result.map_err(|errors| {
-        let errors = exclude_warnings_from_dependencies(errors, root_files);
         noirc_errors::reporter::report_all(
             file_manager.as_file_map(),
             &errors,
@@ -149,7 +145,6 @@ pub fn report_errors<T>(
         )
     })?;
 
-    let errors = exclude_warnings_from_dependencies(errors, root_files);
     noirc_errors::reporter::report_all(
         file_manager.as_file_map(),
         &errors,
@@ -158,20 +153,4 @@ pub fn report_errors<T>(
     );
 
     Ok(t)
-}
-
-fn exclude_warnings_from_dependencies(
-    errors: Vec<FileDiagnostic>,
-    root_files: &HashSet<FileId>,
-) -> Vec<FileDiagnostic> {
-    errors
-        .into_iter()
-        .filter(|error| {
-            if error.diagnostic.is_warning() {
-                root_files.contains(&error.file_id)
-            } else {
-                true
-            }
-        })
-        .collect()
 }
