+++ tooling/nargo_cli/src/cli/debug_cmd.rs
@@ -132,7 +132,6 @@ pub(crate) fn compile_bin_package_for_debugging(
     report_errors(
         compilation_result,
         &workspace_file_manager,
-        &root_files,
         compile_options.deny_warnings,
         compile_options.silence_warnings,
     )
