+++ tooling/nargo_cli/src/cli/export_cmd.rs
@@ -82,12 +82,10 @@ fn compile_exported_functions(
             // TODO: We should to refactor how to deal with compilation errors to avoid this.
             let program = compile_no_check(&mut context, compile_options, function_id, None, false)
                 .map_err(|error| vec![FileDiagnostic::from(error)]);
-            let root_files = context.crate_files(&crate_id);
 
             let program = report_errors(
                 program.map(|program| (program, Vec::new())),
                 file_manager,
-                &root_files,
                 compile_options.deny_warnings,
                 compile_options.silence_warnings,
             )?;
